test(hosted): split RunIntent client coverage
